Run the Docker entrypoint by itself with no shell
The previous non-array syntax invoked a shell to parse the commmand. The
shell remained as process ID 1 and swallowed not just additional
command-line arguments that might be useful to individual deployments,
but also operating system signals such as the one to clean up for a
graceful shutdown. This needless complexity occasionally caused small
inconveniences such as pull request #100.